+++ CodingSession/ViewController.swift
@@ -13,21 +13,11 @@ import Photos
 import RxCocoa
 
 class ViewController: UIViewController {
-    
+
     private let collectionView: UICollectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
-    private lazy var settingsButton: UIButton = {
-        var config = UIButton.Configuration.bordered()
-        config.title = "Go to settings and allow access to videos"
-        
-        let button = UIButton(configuration: config, primaryAction: .init(handler: { _ in
-            UIApplication.shared.open(URL(string: UIApplication.openSettingsURLString)!)
-        }))
-        
-        return button
-    }()
-    
-    lazy var dataSource = GalleryCollectionViewDataSource(collectionView: collectionView, itemSize: UIConstants.itemSize)
-    let viewModel: GalleryViewModel
+    private lazy var settingsButton = UIButton.makeSettingsButton()
+    private lazy var dataSource = GalleryCollectionViewDataSource(collectionView: collectionView, itemSize: UIConstants.itemSize)
+    private let viewModel: GalleryViewModel
     
     private let disposeBag = DisposeBag()
     
@@ -43,17 +33,24 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        setupUI()
+        bind()
+       
+        viewModel.start()
+    }
+    
+    private func setupUI() {
         setupCollectionView()
         setupSettingsButton()
-        
+    }
+    
+    private func bind() {
         viewModel.cellControllers.observe(on: MainScheduler()).subscribe { [weak self] controllers in
             self?.dataSource.set(cellControllers: controllers)
         }
         .disposed(by: disposeBag)
         
         viewModel.settingsButtonHidden.bind(to: self.settingsButton.rx.isHidden).disposed(by: self.disposeBag)
-        
-        viewModel.start()
     }
     
     private func setupSettingsButton() {
@@ -69,7 +66,6 @@ class ViewController: UIViewController {
         collectionView.snp.makeConstraints { make in
             make.edges.equalToSuperview()
         }
-        collectionView.register(ViewControllerCell.self, forCellWithReuseIdentifier: "ViewControllerCell")
         collectionView.delegate = dataSource
         collectionView.dataSource = dataSource
     }
